// Erronka2_WEB/src/app/interfaces/Iikastetxeak.ts
export interface Iikastetxeak {
    CCEN:     number;
    NOM:      string;
    NOME:     string;
    DGENRC:   string;
    DGENRE:   string;
    GENR:     string;
    MUNI:     number;
    DMUNIC:   string;
    DMUNIE:   string;
    DTERRC:   string;
    DTERRE:   string;
    DEPE:     number;
    DTITUC:   string;
    DTITUE:   string;
    DOMI:     string;
    CPOS:     number;
    TEL1:     number;
    TFAX:     number;
    EMAIL:    string;
    PAGINA:   string;
    COOR_X:   string;
    COOR_Y:   string;
    LATITUD:  number;
    LONGITUD: number;
    id:       string;
}